fixed a compiler warning found by Clang 10.0.1 on armv8-a (Samsung S8 aarch64-unknown-linux-android):
./tgeneric.c:208:40: warning: implicit conversion from 'mp_limb_t' (aka 'unsigned long') to 'double' changes value from 18446744073709551615 to 18446744073709551616 [-Wimplicit-int-float-conversion]
		(double) randlimb () / MPFR_LIMB_MAX;
                                     ~ ^~~~~~~~~~~~~
../src/mpfr-impl.h:1277:28: note: expanded from macro 'MPFR_LIMB_MAX'
#define MPFR_LIMB_MAX     ((mp_limb_t) -1)
                           ^~~~~~~~~~~~~~

[From SVN r14098 (trunk)]

+++ tests/tgeneric.c
@@ -205,7 +205,7 @@ test_generic (mpfr_prec_t p0, mpfr_prec_t p1, unsigned int nmax)
                  take a larger input precision in general, else
                  take a smaller precision. */
               xprec *= (prec < 16 ? 256.0 : 1.0) *
-                (double) randlimb () / MPFR_LIMB_MAX;
+                (double) randlimb () / (double) MPFR_LIMB_MAX;
               if (xprec < MPFR_PREC_MIN)
                 xprec = MPFR_PREC_MIN;
             }
